+++ clinicadl/metrics/config/__init__.py
@@ -1,4 +1,4 @@
-from .base import *
+from .base import LossMetricConfig, MetricConfig
 from .classification import *
 from .enum import ImplementedMetric
 from .factory import get_metric_config
