+++ clinicadl/optim/optimizers/__init__.py
@@ -1,3 +1,1 @@
-from .config import OptimizerConfig
-from .enum import ImplementedOptimizer
-from .factory import get_optimizer_config, get_optimizer_from_config
+from .config import ImplementedOptimizer
