+++ clinicadl/optim/optimizers/config/__init__.py
@@ -1,0 +1,2 @@
+from .base import ImplementedOptimizer, OptimizerConfig, get_optimizer_config
+from .configs import *
